Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
PyMuPDF==1.23.8
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4
openai==1.6.1